+++ packages/sdk/src/api/graphql/index.ts
@@ -1,7 +1,0 @@
-import { init } from '@/api/graphql/core.graphql'
-
-export function initGraphQlClients() {
-  init()
-}
-
-export * from './core.graphql'