+++ packages/toolkit/src/composables/index.ts
@@ -1,3 +1,4 @@
+export * from './use-file'
 export * from './use-form'
 export * from './use-form-validation'
 export * from './use-universal-storage'